Adressing code SAT errors
