+++ bundles/org.openhab.binding.sagercaster/src/main/java/org/openhab/binding/sagercaster/internal/handler/ExpiringMap.java
@@ -1,0 +1,49 @@
+package org.openhab.binding.sagercaster.internal.handler;
+
+import java.util.Optional;
+import java.util.SortedMap;
+import java.util.TreeMap;
+
+/**
+ * Copyright (c) 2010-2020 Contributors to the openHAB project
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+import org.eclipse.jdt.annotation.Nullable;
+
+/**
+ * The {@link ExpiringMap} is responsible for storing a list of values of class T
+ * Values older than eldestAge are discarded at each insert of a new one
+ *
+ * @author Gaël L'hopital - Initial contribution
+ */
+class ExpiringMap<T> {
+    private SortedMap<Long, T> values = new TreeMap<>();
+    private @Nullable T agedValue;
+    private long eldestAge = 0;
+
+    public void setObservationPeriod(long eldestAge) {
+        this.eldestAge = eldestAge;
+    }
+
+    public void put(T newValue) {
+        long now = System.currentTimeMillis();
+        values.put(now, newValue);
+        Optional<Long> eldestKey = values.keySet().stream().filter(key -> key < now - eldestAge).findFirst();
+        if (eldestKey.isPresent()) {
+            agedValue = values.get(eldestKey.get());
+            values.entrySet().removeIf(map -> map.getKey() <= eldestKey.get());
+        }
+    }
+
+    public @Nullable T getAgedValue() {
+        return agedValue;
+    }
+}
